requirements.txt
numpy==1.26.2
matplotlib==3.8.1
pandas==2.1.3
scikit-learn==1.3.2
h5py==3.10.0
flake8==6.1.0
tqdm==4.66.1
faiss-cpu==1.7.4
setuptools==60.7.0
jupyterlab==4.0.8